Update and polishing

+++ .zshrc
@@ -10,7 +10,18 @@ plugins=(git git-extras gradle grails jira npm pyhton sudo web-search wd cake co
 
 source $ZSH/oh-my-zsh.sh
 
-export PATH="$HOME/bin:/opt/apache-maven-3.2.5/bin:/opt/jdk1.8.0_11/bin:/opt/gradle-2.0/bin:/usr/local/heroku/bin:/opt/idea-IU-139.224.1/bin:/opt/go_appengine:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/usr/games:/usr/local/games"
+function addp {
+	PATH="$PATH:$1"
+}
+
+addp "$HOME/bin"
+
+addp "/opt/gradle-2.0/bin"
+addp "/opt/apache-maven-3.2.5/bin"
+addp "/opt/jdk1.8.0_25/bin"
+addp "/opt/gradle-2.2.1/bin"
+addp "/opt/idea-IU-139.659.2/bin"
+addp "/opt/go_appengine"
 
 # Fixes UTF8 on Mac OS apparently
 export PYTHONIOENCODING=utf-8
@@ -19,17 +30,20 @@ alias ducks='du -cks * | sort -rn | head'
 alias online='ping -c 3 -i 0.5 -w 3 -q 8.8.8.8 > /dev/null'
 alias fuck='sudo !!'
 
-# Docker and boot2 docker related stuff
+# Docker and boot2docker related stuff
 # DOCKER_HOST may be unstable, let's see how it goes ...
 
 DOCKER_HOST=tcp://192.168.59.103:2376
 DOCKER_CERT_PATH=$HOME/.boot2docker/certs/boot2docker-vm
 DOCKER_TLS_VERIFY=1
 
-docker-enter() {
-	boot2docker ssh '[ -f /var/lib/boot2docker/nsenter ] || docker run --rm -v /var/lib/boot2docker/:/target jpetazzo/nsenter'
-	boot2docker ssh -t sudo /var/lib/boot2docker/docker-enter '$@'
-}
+if [ -n `which docker-enter 2> /dev/null` ]
+	then
+	docker-enter() {
+		boot2docker ssh '[ -f /var/lib/boot2docker/nsenter ] || docker run --rm -v /var/lib/boot2docker/:/target jpetazzo/nsenter'
+		boot2docker ssh -t sudo /var/lib/boot2docker/docker-enter '$@'
+	}
+fi
 
 if [ $(uname) = 'Darwin' ]
 then
@@ -38,5 +52,4 @@ then
 elif [ $(uname) = 'Linux' ]
 then
 	export JAVA_HOME='/opt/jdk1.8.0_31/'
-
 fi
